Fix build on Linux.

+++ src/engine_lib/include/shape/frustum.h
@@ -10,7 +10,7 @@ class ge_frustum {
     ge_frustum() = default;
     ge_frustum(
         const glm::vec3& location, const glm::vec3& forward, const glm::vec3& right, const glm::vec3& up,
-        float near_plane, float far_plane, float fov_deg, float aspect_ratio);
+        float near_plane, float far_plane, unsigned int fov_deg, float aspect_ratio);
 
     // Tests if the AABB is inside of the frustum or intersects it, otherwise returns `false`.
     bool is_aabb_in_frustum(const ge_aabb& aabb_model_space, const glm::mat4& world_mat) const;
